Introduces the Templates class
Primarily, this class reverts some of the changes made to template titles in Gutenberg: https://github.com/WordPress/gutenberg/pull/51428

It keeps some of the ideas with alterations to the singular/plural forms. Revisit this and do a full audit over template names.

This class also sets us up to add some templates that core doesn't yet handle, which need to be added in.