Try a custom Loop template part and area
A Loop template part allows us to quickly interchange the Query Loop block across the entire theme without needing to update it for each top-level template.

Template parts are not without their problems, such as WP adding an extra wrapper element for them, but this has (hopefully) been accounted for in this commit.