Moves breadcrumbs to header patterns.
This change allows us to have a single instance of the Breadcrumbs block that is tied to the user's preferred header. Because of its design, this meant removing the Single Post: Featured template, at least for now.

I'm still not 100% satisfied with the system, but this primarily has more to do with template parts vs. patterns.