Moves comments to a pattern
This change moves comments from a template part to a pattern. It also does away with the comment part and area altogether. This was necessary because WP outputs the wrapper template part HTML markup, even when it is empty.

The change also fixes some of the layout woes I've had dealing with comments.