And even more pattern cleanup.
This commit also includes a fix for global padding and spacing that had long been an issue. Essentially, in the past, I had to use two wrappers inside of several Query Loop patterns so that it would consistently look the same, regardless of where it was used.

That's one less useless `<div>`.