Move single markup back into templates.
WordPress doesn't properly recognize layouts in the post editor when single-type templates pull in a pattern that contains the Post Content block. The root container of the editor is missing the correct classes, such as `.has-global-padding`. When the call to the Post Content block is within the template itself, it works.