First pass at getting the Unravel style variation in.
Loads of changes here where I was really testing the limits of our design system with a dark color theme. It was obvious that there were some places where the system didn't work well that didn't follow that sort of dark text on a light background. This variation let me find the problems and fix (some of) them.