Don't check block registration for enqueueing style.
Since block stylesheets aren't loaded unless the block is in use, this check is unnecessary. But, we also want to support blocks that have previously been inserted but are not registered on the site. The stylesheet will still correctly be loaded in this instance (the registered check blocked this case).