Consolidate Box and Card into a single Card style.
Because these two block styles are so similar, often the same across global style variations, it made sense to merge them.

This change also creates a new Card Footer style to handle card footers when needed. However, this might be better handled as a Card Offset or Card Inner Section (or something) style where we need to offset the background or give top/bottom borders to a nested Group.