Introduce a new `ColorScheme` class.
The goal here is to make the color scheme agnostic from any implementations, such as the toggle button. That way, we can add or switch out the implementation in the future if needed. There's still some architectural work to do that touches on this, but it's mostly outside of the scope.