Introduce `neutral-0` and `neutral-1000` colors.
The goal here is to separate Always Light (White) and Always Dark (Black) colors from flippable light/dark neutrals when the color scheme supports light/dark mode.

There are times when we need these colors to stay the same and times when they should toggle. So we need to account for both scenarios in the color scheme.