Move site header and footer styles to inside the patterns.
Previously, we registered those on the template parts, which seemed like a more ideal solution for a lot of reasons (not really important now, I suppose). Anyway, the styles did not show in the template part editor because they were added through the template at a level above.

Ideally, Core would provide a way to style specific template parts via `theme.json`, but we work with what we got.