Gigantic overhaul of the design system.
I realized some ideas were not working well with WordPress when mixing theme style variations with color style variations. The system with styling `.wp-site-blocks`, which was changed to the `site-container` wrapper was a byproduct of needing a dedicated wrapper container.

It's too much to explain all of the changes. More to come, but this is a baseline for getting a lot of the ideas actually working.